RAG-cap/requirements.txt
langchain
langchain-community
langchain-google-genai
chromadb
sentence-transformers
python-dotenv